Track pending from root
