+++ packages/hash/.size-snapshot.json
@@ -1,7 +1,7 @@
 {
   "dist/hickory-hash.es.js": {
-    "bundled": 7871,
-    "minified": 2801,
+    "bundled": 7545,
+    "minified": 2793,
     "gzipped": 1057,
     "treeshaked": {
       "rollup": {
@@ -14,18 +14,18 @@
     }
   },
   "dist/hickory-hash.js": {
-    "bundled": 8055,
-    "minified": 2977,
-    "gzipped": 1109
+    "bundled": 7715,
+    "minified": 2956,
+    "gzipped": 1105
   },
   "dist/hickory-hash.umd.js": {
-    "bundled": 18313,
-    "minified": 5380,
-    "gzipped": 2153
+    "bundled": 18779,
+    "minified": 5517,
+    "gzipped": 2200
   },
   "dist/hickory-hash.min.js": {
-    "bundled": 18313,
-    "minified": 5380,
-    "gzipped": 2153
+    "bundled": 18779,
+    "minified": 5517,
+    "gzipped": 2200
   }
 }
