+++ packages/root/src/types/locationUtils.ts
@@ -1,8 +1,8 @@
 import { Location, SessionLocation, PartialLocation } from "./location";
 
 export interface QueryFunctions {
-  parse: (query?: string) => Q;
-  stringify: (query?: Q) => string;
+  parse: (query?: string) => any;
+  stringify: (query?: any) => string;
 }
 
 export type RawPathname = (pathname: string) => string;
