+++ packages/root/types/navigationHandler.d.ts
@@ -1,0 +1,7 @@
+import { ResponseHandler, PendingNavigation } from "./types/navigation";
+export default function responder(): {
+    emitNavigation: (nav: PendingNavigation) => void;
+    clearPending: () => void;
+    cancelPending: (action?: "push" | "replace" | "pop" | undefined) => void;
+    setHandler: (fn: ResponseHandler) => void;
+};
